data_pipeline/dev_requirements.txt
apache-airflow==2.3.4
pendulum==2.1.2
feast[redis]==0.24.0
pandas==1.4.4
pyspark==3.0.1